aufs: xino 1/2, core
XINO and XIB files are to maintain the inode numbers in aufs
(cf. struct.txt and aufs manual in aufs-util.git).

The maximum file size of XINO is limited to LLONG_MAX or the s_maxbytes
in filesystem's superblock where XINO files are placed (s_maxbytes may
be smaller than LLONG_MAX). So the support-able largest inode number on
a branch is less than LLONG_MAX/4-1.
This is the current limitation of aufs.
On 64bit environment, this limitation becomes more strict and the
supported largest inode number is less than LLONG_MAX/8-1.
You can confirm its size via debugfs (in later commit).

Sometimes the size of XINO file can be a problem, ie. too big,
particularly for tmpfs. In this case, another separatte patch
tmpfs-ino.patch in aufs4-standalone.git is recommended (as well as
vfs-ino.patch). The patch makes tmpfs to maintain inode number within
itself and supress its discontiguous distribution.

See also the document in this commit.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>