aufs: xino 2/2
XINO and XIB files are read and written frequently after unlinked, and
it means that the remote filesystems are not suitable for
them. Additionally aufs shows their metadata via debugfs (in later
commit). To make it easier to do this, aufs expects branchfs to maintain
their i_size and i_blocks. And it means some filesystem are not suitable
for XINO.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>