aufs: sysaufs
The branch path can be much longer and it is not suitable to print via
/proc/mounts as a part of mount options. Aufs can show it either
separately via sysfs or /proc/mounts (as a part of mount options).
This approach affects the lifetime of aufs objects and sbinfo contains
kobject (in another commit). Theoritically user can disable
CONFIG_SYSFS, but the lifetime management is always necessary. So
supporting sysfs is splitted into two files, sysaufs.c and sysfs.c.
sysaufs.c is always compiled, but sysfs.c is compiled only when
CONFIG_SYSFS is enabled.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>