aufs: si_pid bitmap/tree
In freeing aufs iinfo objects, it acquires the internal rw_sem (see
another commit in detail). Since iinfo can be freed anytime, a deadlock
may happen due to the rw_sem. To prevent this problem, this commit
introduces an internal bitmap and a tree of PID who already acquires the
rw_sem.
This is another (very) ugly approach which I don't like.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>