aufs: hnotify 2/3, body
The feature is constructed by two layers. One is generic interface, and
the other is exact implementation. This is rather historical. Originally
aufs implemented this feature based upon 'inotify.' Later 'fsnotify'
made 'inotify' obsolete. During the transition period, these two layers
were introduced to support both of 'inotify' and 'fsnotify.' Currently
only 'fsnotify' is supported, but the layers are kept for the future.

This feature is compiled only when CONFIG_AUFS_HNOTIFY is enabled.
See also the document in previous commit.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>