aufs: export via NFS 2/2
The main part is in previous commit.
This commit handles the generation of aufs objects, to make sure the
inode in the file handle is still valid.
In orer not to confuse NFSD, the various operaion returns ESTALE for
NFSD where it used to return EBUSY.

See also the document in this commit.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>