aufs: export via NFS 1/2, body
Implement exporting via NFS.
The file handle is rather large (40 bytes at most + the file handle on a
branch).
The non-virtual filesystems can use an anonymous (disconnected) dentry
as long as the inode is identifed, but aufs needs a dentry with dinfo
which is usually contructed. So aufs has to find or generate the normal
dentry from the file handle in decoding.
If both of the dentry and the inode which are correspondging the file
handle are still in cache, then they are returnd immediately. Otherwise
aufs has to find the cached parent dir from the file handle. If the
parent dir is not cached either, the aufs tries
- decode the branch fs's file handle
- generate the path of the parent dir on the branch
- convert the branch path to aufs's path
- lookup the inode number under the aufs's path
The last one is the slowest case.

See also the document in later commit.
This is compiled only when CONFIG_AUFS_EXPORT is enabled.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>