aufs: xino truncation
As mentioned earlier, sometimes the size of XINO file is a problem.
Aufs has a feature to truncate it. But it may not be so effective in
some cases, and you may want to stop discontiguous distribution of the
inode numbers on branch fs.
See also the log in another commit.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>