aufs: hnotify 1/3, headers
This is the hardest test to support UDBA (users' direct branch access).
It uses 'fsnotify' internally, detect UDBA, decrements the generation of
the cached aufs objects. At the next access to the file, aufs detects
the generation is obsoleted and tries refreshing it. Finally aufs cache
will be updated to latest status.

The fsnotify is set on the cached dirs on the non-RR branches.
The RR (real readonly) branches will never be modified and it is
unnecessary to set fsnotify for them.

This commit is for the declarations mainly, and the body parts will be
in succeeding commits.

This feature is compiled only when CONFIG_AUFS_HNOTIFY is enabled.
See also the document in this commit.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>