aufs: dirwh option
This is a feature to optimize for rmdir and rename dir.
When the number of whiteouts under the target dir is very many, it may
take a long time to remove them all. To prevent this, 'dirwh=%d' option
specfies the watermark to decide when to remove them.

For details, see aufs manual in aufs-util.git.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>