aufs: atomic_open 5/5, the body aufs_atomic_open()
->atomic_open() is another monster (the other is ->rename() of cource).
It operates look-up, create, and open in a single method.
Strictly speaking the behaviour is not atomic from the branch fs's point
of view, while the atomicity is kept in aufs's. This is a second-best
approach. For details, refer to the design doccument in previous commit.

A simple list 'si_aopen' can be put into aufs inode object, but I don't
think it a good idea because the number of inodes is much larger than
the number of super_block. If I put it into inode, it can be an
un-interesting memory pressure I am afraid.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>
(cherry picked from commit 6ddaaeb205053424c357aa659987ab8d479fb6c9)