aufs: for linux-4.2-rc1, symlink 1/3, a new list per sb
In linux-4.2-rc1 there is massive changes around symbolic link.
The most important change is the parameter cookie which is a key between
VFS and FS. As a middle layer between them, aufs keeps the cookie
internally and convert it for the real FS.
Note that aufs supports the dynamic branch manipulation. If it happens
between follow_link() and put_link() calls, the topmost h_inode may not
be identical. In order to support such situation, aufs also stores
h_inode in the list.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>