aufs: bugfix, suppress VFS delayed fput in copy-up
For linux-3.6 there was a commit
	4a9d4b0 2012-07-22 switch fput to task_work_add
which introduced the delayed fput. It is good for performance, but for
aufs internal copy-up, a problem was born.

The newly copied-up file is opened to be written and its inode's
i_writecount is incremented (via get_write_access()). On the other hand,
execve() rejects when the executable file inode's i_writecount is
incremented. It is "actual" fput that decrements i_writecount.

Introducing the delayed fput created a small window for execve() to
return ETXTBSY. The copy-up is done by aufs workqueue kthread and fput()
call from copy-up is postponed by the delayed fput feature. So before
the actual fput is done, user could not run a just copied-up executable.

This commit calls __fput_sync() which is a variation of fput() and put
the file synchoronously, in order to make it possible to execeve() just
after copy-up.

Reported-by: James Burry <jburry@gmail.com>
See-also: http://www.mail-archive.com/aufs-users@lists.sourceforge.net/msg05257.html
Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>
(cherry picked from commit 8e80c53941c66d026fbb9bd88b6222b06163344c)