aufs: minor, shrinkable krealloc
The original krealloc function in linux kernel never shrink the given
memory region. It is good generally, but not for aufs' branch
management.
When a user adds very many branches and then removes most of them, then
the large memory left as consumed for aufs' internal pointer arrays.

A simple wrapper for krealloc() in this commit, if possible, tries
allocating a new smaller memory region, copying the contents from the
old region to the new allocated one, and free the old one.
The decision is made based upon kmalloc_index(). But when CONFIG_SLOB is
enabled, kmalloc_index() is not usable. For SLOB, this wrapper always
calls the original krealloc(). I don't know it is shrinkable.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>