aufs: bugfix, update_time() for the special files
i_op->update_time() was introduced in linux-v3.5, and aufs3.5
implemented it. But that was incorrect and aufs might try updating the
inode attributes on its readonly branch, particulary the inode was a
special file. Obviously updating on the readonly branch is a bad
violation of aufs policy.

This commit prohibits updating on the readonly branch, but this is
another violation of an ordinary behaviour of VFS. The special files may
exist on the readonly branch even if it was opened for writing. In this
case, the timestamps of the special file won't be updated.
For such case, if this commit copied it up when its ->update_time() is
called, then the communication channel would be corrupted. Copying-up
and setting the timestamps are less important than keeping the
communication I believe.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>