aufs: convert pid bitmap to a flag per task
Anciently, aufs2.0 for linux-v2.6.27, aufs_sbinfo introduced a radix
tree in order to make aufs possible to free aufs_iinfo anytime. The
bitmap tracks the pid which holds aufs internal rwsem (see the old
commits below), and aufs3 and aufs4 inherit it. The tree was converted
to a bitmap later in aufs4.
But as written in the original commit, I don't like such approach. Now
aufs decides borrowing a room for that from task_struct. I don't like
this approach either, but it is a little better since it should be a
little faster and smaller.

aufs4
f840089 2016-03-21 aufs: minor, convert radix_tree for si_pid to bitmap array
9e05994 2015-03-30 aufs: si_pid bitmap/tree

aufs3 (for linux-v3.0)
f960a1f 2011-08-10 aufs: super_block, from aufs2.2-3.0

aufs2.0 (for linux-v2.6.27)
89d5929 2010-06-16 aufs: possible bugfix, sbinfo lock in deleting inode, si_pid functions

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>