aufs: fine-grained xino lock 2/2, caller
Replace a rather larger critical section by a mutex by a new
fine-grained logical lock per inode-number.
It was expected to be a little help for the performance, but it is not
so much (based on my small tests). Because the lock contention by this
mutex was not so common. The fine-grained lock is not a help for the
non-contended lock. But I hope it will appear better on the much larger
system under a heavy workload.

The inode numbers in the new array means that aufs is assigning the
inode number for that and means "logically locked."

See also the commit in aufs2-2.6.git#aufs2.2-35 branch (which is for
linux-v2.6.35)
9f8ad8c 2010-06-03 aufs: possible bugfix, revalidate inode race between readdir and lookup

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>