aufs: dirren 1/6, inum-list of the renamed dir in a branch
This commits brings a list of the inode numbers which indicates the
logically renamed dir into a branch. The list will be refered in lookup,
and its lifetime is equivalent to the branch's, ie. the list is
loaded/created in adding a branch, and stored/deleted in deleting a
branch. The simple storing happens in remouting and unmounting aufs too.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>