aufs: bugfix, fanotify
Linux FANOTIFY feature is rather exotic.
During the monitoring, if a new file is created, fanotify tells it to
the user. In this notification, fanotify opens the created
file internally.
The most interest thing is it happens before the creator process
completes open(2). So both of the listner and the creator enters open(2)
concurrently.
It caused a deadlock in aufs since aufs aquires a lock for that file,
which may be handled by aufs copyup-on-open or moveup-on-open feature
internally.

This commit shrinks the lock region and allows concurrent open(2).

Reported-by: Ariel Zelivansky <ariel@twistlock.com>
See-also: https://bugzilla.kernel.org/show_bug.cgi?id=196815
Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>