aufs: for v4.10, XINO(read) handles EINTR from the dying process
The commit in linux-v4.10-rc7
	5abf186 2017-02-03 mm, fs: check for fatal signals in do_generic_file_read()
will make aufs hung-up under a certain condition.
The condition is (roughly)
- linux-v4.10-rc7 and later.
- aufs has many files (over 32K inode-numbers are assigned).
- the process is killed by SIGKILL.
- the dying process held a unliked file whose inode-number is cold.
  here "cold" means out of aufs internal XINO bitmap cache. so aufs has
  to read XINO bitmap from a disk in order to maintain the bitmap.
- the fs where XINO files are placed calls do_generic_file_read() to
  imprelent its ->read().

In linux-v4.3, the was a very similar situation by the commit
        296291c 2015-10-23 mm: make sendfile(2) killable
but it was "write to XINO" instead of "read from XINO" case.
And it was fixed by the commit
	5e439ff 2016-01-05 aufs: for 4.3, XINO handles EINTR from the dying process
which made aufs4.3 retries "write to XINO" in another context after
EINTR.

Here the same approach is applied to the case of "read from XINO."

Reported-by: Eddie Horng <eddiehorng.tw@gmail.com>
Tested-by: Eddie Horng <eddiehorng.tw@gmail.com>
Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>
(cherry picked from commit f2474d8863aaf2203052a717b72b987dd833a78c)