aufs: refine xino, minor, remove an unused internal member
The purpose of 'old' member in this internal structure was the debugging
and to be revertible on error in aufs1 era.  But the debugging test was
removed when aufs evolved to aufs2, and being revertible seems overspec
now.  The member might have to be removed in aufs2 era I am afraid.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>