aufs: xinoarray 1/4, a new member sbinfo->si_ximaxent
It represents the max entries in a xino file, based on the calculation
of s_maxbytes.  When a new branch fs' inode comes, aufs decides the
index in the xinoarray based on the calculation using this value.
Honestly speaking, I am afraid that the branch fs may not set s_maxbytes
accurately though.
If you want to test xinoarray, then you can set a small number to it.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>