aufs: xinoarray 3/4, make it indexable
Use xi_file array actually which was created in previous commit.
The important part in this commit is adding 'idx' parameter to
au_xino_file() and au_xino_file_set().
This commit is just a part of 'xinoarray' series and doesn't create
multiple XINO files in a single struct au_xino yet.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>