aufs: minor, use div64_u64_rem()
Some versions of GCC for 32bit arch generate __divmoddi4 call for the
64bit division and modulo operator.  libgcc defines __divmoddi4 function
and it is resolved at the link-time.  But some versions of libgcc don't
provide the function and cause a linkage-error in building aufs module.
I am afraid it is a problem of LIBGCC.  Here we have a workaround for
it.

Reported-by: peabee as a github user
See-also: https://github.com/sfjro/aufs4-standalone/commit/df40a305a40d9e30c3d5f1c0385d52be4dd73453#commitcomment-30298763
Tested-by: peabee as a github user
Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>