aufs: possible bugfix, remove sbinfo->si_xino_brid
The purpose of this member is a small optimization, which finds the
branch index where the XINO files are placed at branch root dir.
Maintaining this member costs high and confuses me in the branch-mgmt.
Moreover it requires down_write(si_rwsem) in filling an asynchronously
creating xino-array element.  That is bad.
Let's remove it and switch to the simpler uncached approach.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>