aufs: temp workaround for atomic_open on nfs4 branch
When the RW branch is nfs4, there is a case of a new aufs file doesn't
appear in the result of readdir(3) just after the creation.  I don't
know why.  here is a workaround for it which makes the parent dir to
be re-validated at the next access time.  It may not be a perfect
solution and it may just reduce the chances of the problem I am afraid.
But I have no other solution.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>