aufs: minor, keep br_count positive
In au_reopen_nondir(), the ref-count of the branch is decremented
temporary by au_set_h_fptr(,,NULL).  This is not a big problem nor a
bug, since the counter is inremented later in the same function.
But I guess keeping it positive is better for the cleaner logic.
The behaviour should not change.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>