aufs: br_nfiles and br_count 1/2, minor, br_nfiles can allow deleting a branch
The heart of this series is in succeeding commit.
This commit simply introcudces a new counter br_nfiles in struct
au_branch. The new counter represents the number of hidden files opened
by aufs. Before this commit, br_count represented the same thing, but we
need a diffrent counter to represent that the branch is busy by other
than opened file. These two counters should not be confused.

This commit is to extract the simple meaning (the number of opened
files) from br_count.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>